@@ -1,3 +1,3 @@
 module github.com/twmb/go-sliceheap
 
-go 1.12
+go 1.20
