@@ -9,7 +9,7 @@ import (
 
 func ExampleHeap() {
 	// inner shows that we can create heap in one function and return it.
-	inner := func() sliceheap.Heap {
+	inner := func() sliceheap.Heap[int] {
 		a := []int{3, 2, 4, 5, 1, 0, 6}
 		h := sliceheap.On(&a, func(i, j int) bool { return a[i] > a[j] })
 		heap.Init(h)
@@ -17,7 +17,7 @@ func ExampleHeap() {
 	}
 	h := inner()
 	// We can see the heap sort itself by checking the backing slice.
-	fmt.Println(h.View().([]int))
+	fmt.Println(h.View())
 	// Push a few more elements.
 	heap.Push(h, 8)
 
@@ -25,7 +25,7 @@ func ExampleHeap() {
 	// pointer to the slice. This is only necessary if working on a slice
 	// that you did not pass directly to sliceheap.On, i.e., you have lost
 	// the original pointer you used.
-	ptr := h.Pointer().(*[]int)
+	ptr := h.Pointer()
 	fmt.Println(*ptr)
 	heap.Push(h, 7)
 	fmt.Println(*ptr)
@@ -59,16 +59,4 @@ func ExampleOn() {
 	fmt.Println(heap.Pop(h).(int))
 	// Output:
 	// 6
-}
-
-func ExampleHeap_Peek() {
-	a := []int{3, 1, 2}
-	h := sliceheap.On(&a, func(i, j int) bool { return a[i] < a[j] })
-	*h.Peek().(*int) = 4 // modify the first element
-	fmt.Println(a)
-	heap.Fix(h, 0) // and then fix its position
-	fmt.Println(a)
-	// Output:
-	// [4 1 2]
-	// [1 4 2]
 }
